# backend/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Authentication API",
    "description": "API for user authentication",
    "version": "1.0.0"
  },
  "paths": {
    "/auth/login": {
      "post": {
        "tags": ["General"],
        "summary": "Authenticate user and get a JWT token",
        "description": "Login with your Google email to receive a JWT token",
        "operationId": "login",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "User email for authentication",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LoginRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully logged in, returns JWT token.",
            "schema": {
              "$ref": "#/definitions/LoginResponse"
            }
          },
          "400": {
            "description": "Invalid email"
          }
        }
      }
    },
    "/refresh": {
      "post": {
        "tags": ["General"],
        "summary": "Refresh JWT token",
        "description": "Provide a valid refresh token to receive a new access token",
        "operationId": "refreshToken",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Refresh token request",
            "required": true,
            "schema": {
              "$ref": "#/definitions/RefreshTokenRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully refreshed token, returns new JWT token.",
            "schema": {
              "$ref": "#/definitions/RefreshTokenResponse"
            }
          },
          "401": {
            "description": "Unauthorized if no refresh token is provided"
          },
          "403": {
            "description": "Forbidden if refresh token is invalid"
          }
        }
      }
    },   
...
    "/logout": {
      "get": {
        "tags": ["General"],
        "summary": "Log out the user and clear session",
        "description": "Logs out the user by destroying the session and clearing the JWT cookie.",
        "operationId": "logout",
        "responses": {
          "200": {
            "description": "Successfully logged out."
          },
          "400": {
            "description": "Failed to log out."
          }
        }
      }
    },
    "/companies": {
      "get": {
        "tags": ["Companies"],
        "summary": "Get all companies",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of all companies",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Company"
              }
            }
          },
          "500": {
            "description": "Error fetching companies"
          }
        }
      },
      "post": {
        "tags": ["Companies"],
        "summary": "Create a new company",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "company",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NewCompany"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Company created successfully",
            "schema": {
              "$ref": "#/definitions/Company"
            }
          },
          "400": {
            "description": "Missing required fields (name, address)"
          },
          "500": {
            "description": "Error creating company"
          }
        }
      },
      "options": {
        "tags": ["Companies"],
        "summary": "Get allowed methods for companies",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of allowed methods for the companies route"
          }
        }
      }
    },
...
    "/jobs": {
      "get": {
        "tags": ["Jobs"],
        "summary": "Get all jobs",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of all jobs",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Job"
              }
            }
          },
          "500": {
            "description": "Error fetching jobs"
          }
        }
      },
      "post": {
        "tags": ["Jobs"],
        "summary": "Create a new job",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "job",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NewJob"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Job created successfully",
            "schema": {
              "$ref": "#/definitions/Job"
            }
          },
          "400": {
            "description": "Missing required fields (title, description)"
          },
          "500": {
            "description": "Error creating job"
          }
        }
      }
    },
...
    "/applications": {
      "get": {
        "tags": ["Applications"],
        "summary": "Fetch all applications",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of all applications",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Application"
              }
            }
          },
          "500": {
            "description": "Error fetching applications"
          }
        }
      },
      "post": {
        "tags": ["Applications"],
        "summary": "Create a new application",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{
          "in": "body",
          "name": "application",
          "description": "New application data",
          "schema": {
            "$ref": "#/definitions/ApplicationCreateRequest"
          }
        }],
        "responses": {
          "201": {
            "description": "Application created successfully",
            "schema": {
              "$ref": "#/definitions/Application"
            }
          },
          "500": {
            "description": "Error creating application"
          }
        }
      },
      "options": {
        "tags": ["Applications"],
        "summary": "Fetch allowed methods for applications",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Allowed HTTP methods",
            "schema": {
              "$ref": "#/definitions/AllowedMethodsResponse"
            }
          }
        }
      }
    },
...
    "/users": {
      "get": {
        "tags": ["Users"],
        "summary": "Fetch all users",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "A list of users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/User"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "tags": ["Users"],
        "summary": "Create a new user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [{
          "in": "body",
          "name": "user",
          "description": "New auser data",
          "schema": {
            "$ref": "#/definitions/UserCreateRequest"
          }
        }],
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid email format"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "options": {
        "tags": ["Users"],
        "summary": "List allowed methods",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Allowed methods",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "allowed_methods": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "LoginRequest": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "example": "user@gmail.com"
        }
      }
    },
    "LoginResponse": {
      "type": "object",
      "properties": {
        "token": {
          "type": "string",
          "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VybmFtZSI6ImFkbWluIiwiaWF0IjoxNzM2MTU5MjA2LCJleHAiOjE3MzYxNjI4MDZ9.Ovm18jYu4pUbkBe4pKKYue0rSIpr0SB7mczsCa9R4G0"
        }
      }
    },
    "RefreshTokenRequest": {
      "type": "object",
      "properties": {
        "refreshToken": {
          "type": "string",
          "example": "dGhpc2lzYXJlZnJlc2h0b2tlbg=="
        }
      }
    },
    "RefreshTokenResponse": {
      "type": "object",
      "properties": {
        "accessToken": {
          "type": "string",
          "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VybmFtZSI6ImFkbWluIiwiaWF0IjoxNzM2MTU5MjA2LCJleHAiOjE3MzYxNjI4MDZ9.Ovm18jYu4pUbkBe4pKKYue0rSIpr0SB7mczsCa9R4G0"
        }
      }
    },
    "Company": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "sector": {
          "type": "string"
        },
        "contact_info": {
          "type": "object"
        },
        "address": {
          "type": "string"
        },
        "services": {
          "type": "object"
        },
        "created_at": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "NewCompany": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "address": {
          "type": "string"
        },
        "sector": {
          "type": "string"
        },
        "contact_info": {
          "type": "object"
        },
        "services": {
          "type": "object"
        }
      },
      "required": ["name", "address"]
    },
...
    "Job": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "example": 1
        },
        "title": {
          "type": "string",
          "example": "Software Engineer"
        },
        "description": {
          "type": "string",
          "example": "Develop and maintain software applications."
        },
        "company_id": {
          "type": "integer",
          "example": 1
        }
      }
    },
    "NewJob": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "company_id": {
          "type": "integer"
        },
        "type": {
          "type": "string"
        }
      }
    },
...
    "Application": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "user_id": {
          "type": "integer"
        },
        "job_id": {
          "type": "integer"
        },
        "application_date": {
          "type": "string",
          "format": "date-time"
        },
        "status": {
          "type": "string",
          "enum": ["pending", "accepted", "rejected"]
        }
      }
    },
    "ApplicationCreateRequest": {
      "type": "object",
      "properties": {
        "user_id": {
          "type": "integer"
        },
        "job_id": {
          "type": "integer"
        }
      },
      "required": ["user_id", "job_id"]
    },
...
    "AllowedMethodsResponse": {
      "type": "object",
      "properties": {
        "allowed_methods": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "example": ["GET", "POST", "OPTIONS"]
        }
      }
    },
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "username": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "role": {
          "type": "string",
          "enum": ["admin", "user", "guest"]
        },
        "created_at": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "UserCreateRequest": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        },
        "email": {
          "type": "string",
          "format": "email"
        },
        "password": {
          "type": "string"
        },
        "role": {
          "type": "string",
          "enum": ["admin", "user", "guest"]
        }
      },
      "required": ["username", "email", "password"]
    },
...
  },
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Enter your Bearer token in the format: Bearer token",
      "x-display-name": "Bearer Token",
      "x-token-prefix": "Bearer "
    }
  }
}